refactor(oracle): extract generator sampling primitives module
