+++ tools/oracle/engine/generator.py
@@ -8,26 +8,39 @@ import ipaddress
 import random
 import sys
 from collections.abc import Mapping, Sequence
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import TypeVar
 
 from .model import JSONObject, PacketPlan
+from .sampling import (
+    EPHEMERAL_PORT_MAX,
+    EPHEMERAL_PORT_MIN,
+    T,
+    _SKIP_FIELD,
+    _SamplingContext,
+    _SkipField,
+    _different_ipv4,
+    _different_ipv6,
+    _different_mac,
+    _different_port,
+    _identifier_part,
+    _integer_domain_value,
+    _IPV4_DOCUMENTATION_NETWORKS,
+    _IPV6_DOCUMENTATION_NETWORK,
+    _mac_for_domain,
+    bounded_int,
+    byte_payload,
+    documentation_ipv4,
+    documentation_ipv6,
+    documentation_mac,
+    ephemeral_port,
+    plan_identifier,
+    weighted_choice,
+)
 from .spec_loader import load_oracle_specs
 
 
-T = TypeVar("T")
-
-EPHEMERAL_PORT_MIN = 49152
-EPHEMERAL_PORT_MAX = 65535
 SUPPORTED_LAYER_BACKEND = "libcrafter"
 
-_IPV4_DOCUMENTATION_NETWORKS = (
-    ipaddress.IPv4Network("192.0.2.0/24"),
-    ipaddress.IPv4Network("198.51.100.0/24"),
-    ipaddress.IPv4Network("203.0.113.0/24"),
-)
-_IPV6_DOCUMENTATION_NETWORK = ipaddress.IPv6Network("2001:db8::/32")
 _DERIVED_DOMAINS = {"derived"}
 _IGMP_DOC_GROUP = "233.252.0.17"
 _IGMP_DOC_GROUP_ALT = "233.252.0.61"
@@ -407,112 +420,6 @@ _SCAPY_MATERIALIZED_LAYERS = {
     "zigbee_aps",
     "zigbee_nwk",
 }
-
-
-def weighted_choice(rng: random.Random, choices: Sequence[tuple[T, int]]) -> T:
-    """Choose one item from integer weights using only the supplied RNG."""
-
-    if not choices:
-        raise ValueError("weighted choice requires at least one option")
-
-    total = 0
-    for _, weight in choices:
-        if weight < 0:
-            raise ValueError(f"weights must be non-negative: {weight}")
-        total += weight
-    if total <= 0:
-        raise ValueError("weighted choice requires a positive total weight")
-
-    cursor = rng.randrange(total)
-    seen = 0
-    for value, weight in choices:
-        seen += weight
-        if cursor < seen:
-            return value
-
-    raise RuntimeError("weighted choice cursor escaped total weight")
-
-
-def bounded_int(rng: random.Random, minimum: int, maximum: int) -> int:
-    """Return an inclusive integer in the configured bounds."""
-
-    if minimum > maximum:
-        raise ValueError(f"minimum cannot exceed maximum: {minimum} > {maximum}")
-    return rng.randint(minimum, maximum)
-
-
-def byte_payload(
-    rng: random.Random,
-    *,
-    min_length: int = 0,
-    max_length: int = 64,
-) -> bytes:
-    """Generate a deterministic byte payload with a deterministic length."""
-
-    length = bounded_int(rng, min_length, max_length)
-    return bytes(rng.randrange(256) for _ in range(length))
-
-
-def documentation_ipv4(rng: random.Random) -> str:
-    """Select an IPv4 address from RFC 5737 documentation networks."""
-
-    network = weighted_choice(
-        rng,
-        tuple((network, 1) for network in _IPV4_DOCUMENTATION_NETWORKS),
-    )
-    host = bounded_int(rng, 1, network.num_addresses - 2)
-    return str(network.network_address + host)
-
-
-def documentation_ipv6(rng: random.Random) -> str:
-    """Select an IPv6 address from the RFC 3849 documentation prefix."""
-
-    host = bounded_int(rng, 1, (1 << 96) - 2)
-    return str(ipaddress.IPv6Address(int(_IPV6_DOCUMENTATION_NETWORK.network_address) + host))
-
-
-def documentation_mac(rng: random.Random) -> str:
-    """Select an EUI-48 documentation address from the RFC 7042 block."""
-
-    suffix = bounded_int(rng, 0, 255)
-    return f"00:00:5e:00:53:{suffix:02x}"
-
-
-def ephemeral_port(rng: random.Random) -> int:
-    """Select an IANA dynamic/private port."""
-
-    return bounded_int(rng, EPHEMERAL_PORT_MIN, EPHEMERAL_PORT_MAX)
-
-
-def plan_identifier(
-    *,
-    seed: int,
-    profile: str,
-    index: int,
-    root: str,
-    stack_name: str,
-    family: str,
-    case: str | None = None,
-    feature: str | None = None,
-) -> str:
-    """Build a stable packet plan identifier with reproduction coordinates."""
-
-    subject_kind = "case" if case is not None else "feature"
-    subject = case if case is not None else feature
-    if subject is None:
-        subject_kind = "case"
-        subject = "default"
-    return "/".join(
-        (
-            f"seed-{seed}",
-            f"profile-{profile}",
-            f"index-{index:06d}",
-            f"root-{_identifier_part(root)}",
-            f"stack-{_identifier_part(stack_name)}",
-            f"family-{family}",
-            f"{subject_kind}-{_identifier_part(subject)}",
-        )
-    )
 
 
 def load_stack_grammar(path: str | Path | None = None) -> JSONObject:
@@ -2546,107 +2453,6 @@ class PacketGenerator:
         return weighted_choice(ctx.rng, choices)
 
 
-class _SkipField:
-    pass
-
-
-_SKIP_FIELD = _SkipField()
-
-
-@dataclass(slots=True)
-class _SamplingContext:
-    rng: random.Random
-    profile: str
-    feature_weights: Mapping[str, int]
-    stack: list[str]
-    payload_min: int
-    payload_max: int
-    feature: str | None
-    case: str
-    sampled_layers: dict[str, JSONObject] = field(default_factory=dict)
-    _payload: bytes | None = None
-    _src_mac: str | None = None
-    _dst_mac: str | None = None
-    _src_port: int | None = None
-    _dst_port: int | None = None
-    _src_ipv4: str | None = None
-    _dst_ipv4: str | None = None
-    _src_ipv6: str | None = None
-    _dst_ipv6: str | None = None
-    _arp_sender_ip: str | None = None
-    _arp_target_ip: str | None = None
-
-    @property
-    def payload(self) -> bytes:
-        if self._payload is None:
-            self._payload = byte_payload(
-                self.rng,
-                min_length=self.payload_min,
-                max_length=self.payload_max,
-            )
-        return self._payload
-
-    @property
-    def src_mac(self) -> str:
-        if self._src_mac is None:
-            self._src_mac = documentation_mac(self.rng)
-        return self._src_mac
-
-    @property
-    def dst_mac(self) -> str:
-        if self._dst_mac is None:
-            self._dst_mac = _different_mac(self.rng, self.src_mac)
-        return self._dst_mac
-
-    @property
-    def src_port(self) -> int:
-        if self._src_port is None:
-            self._src_port = ephemeral_port(self.rng)
-        return self._src_port
-
-    @property
-    def dst_port(self) -> int:
-        if self._dst_port is None:
-            self._dst_port = _different_port(self.rng, self.src_port)
-        return self._dst_port
-
-    @property
-    def src_ipv4(self) -> str:
-        if self._src_ipv4 is None:
-            self._src_ipv4 = documentation_ipv4(self.rng)
-        return self._src_ipv4
-
-    @property
-    def dst_ipv4(self) -> str:
-        if self._dst_ipv4 is None:
-            self._dst_ipv4 = _different_ipv4(self.rng, self.src_ipv4)
-        return self._dst_ipv4
-
-    @property
-    def src_ipv6(self) -> str:
-        if self._src_ipv6 is None:
-            self._src_ipv6 = documentation_ipv6(self.rng)
-        return self._src_ipv6
-
-    @property
-    def dst_ipv6(self) -> str:
-        if self._dst_ipv6 is None:
-            self._dst_ipv6 = _different_ipv6(self.rng, self.src_ipv6)
-        return self._dst_ipv6
-
-    @property
-    def arp_sender_ip(self) -> str:
-        if self._arp_sender_ip is None:
-            self._arp_sender_ip = documentation_ipv4(self.rng)
-        return self._arp_sender_ip
-
-    @property
-    def arp_target_ip(self) -> str:
-        if self._arp_target_ip is None:
-            self._arp_target_ip = _different_ipv4(self.rng, self.arp_sender_ip)
-        return self._arp_target_ip
-
-
 def _field_specs(spec: JSONObject, layer: str) -> list[JSONObject]:
     raw_fields = spec.get("fields")
     if not isinstance(raw_fields, Sequence) or isinstance(raw_fields, (str, bytes)):
@@ -2727,47 +2533,6 @@ def _field_bits(field_spec: JSONObject) -> int:
     if field_type.startswith("uint"):
         return int(field_type.removeprefix("uint"))
     return 16
-
-
-def _integer_domain_value(
-    ctx: _SamplingContext,
-    domain: object,
-    field_name: str,
-    *,
-    bits: int,
-) -> int:
-    maximum = (1 << bits) - 1
-    if isinstance(domain, bool):
-        return int(domain)
-    if isinstance(domain, int):
-        return domain
-    if domain == "boundary":
-        return weighted_choice(ctx.rng, ((0, 1), (maximum, 1)))
-    if domain == "deterministic":
-        return bounded_int(ctx.rng, 0, maximum)
-    if domain == "dynamic":
-        return ephemeral_port(ctx.rng)
-    if domain == "http":
-        return 80
-    if domain == "https":
-        return 443
-    if domain == "dns_client":
-        return ephemeral_port(ctx.rng)
-    if domain == "dns_server":
-        return 53
-    if domain == "bootpc":
-        return 68
-    if domain == "bootps":
-        return 67
-    raise ValueError(f"spec error: unsupported integer domain for {field_name}: {domain!r}")
-
-
-def _mac_for_domain(ctx: _SamplingContext, domain: object, default: str) -> str:
-    if domain == "broadcast":
-        return "ff:ff:ff:ff:ff:ff"
-    if domain == "zero":
-        return "00:00:00:00:00:00"
-    return default
 
 
 def _is_ipv4_root_dhcp_stack(stack: Sequence[str]) -> bool:
@@ -6185,13 +5950,6 @@ def _derive_deck_seed(seed: int, profile: str, names: Sequence[str]) -> int:
     return int.from_bytes(hashlib.sha256(material).digest()[:16], byteorder="big")
 
 
-def _identifier_part(value: str) -> str:
-    output = []
-    for char in value.lower():
-        output.append(char if char.isalnum() else "-")
-    return "".join(output).strip("-") or "default"
-
-
 def _json_object(value: object) -> JSONObject:
     if not isinstance(value, Mapping):
         raise ValueError("expected a JSON object")
@@ -6393,44 +6151,6 @@ def _next_layer_after(stack: Sequence[str], layer: str) -> str | None:
         if next_layer != "payload":
             return next_layer
     return "payload" if "payload" in stack[index + 1 :] else None
-
-
-def _different_mac(rng: random.Random, first: str) -> str:
-    second = documentation_mac(rng)
-    if second != first:
-        return second
-    suffix = (int(first.rsplit(":", 1)[1], 16) + 1) % 256
-    return f"00:00:5e:00:53:{suffix:02x}"
-
-
-def _different_port(rng: random.Random, first: int) -> int:
-    second = ephemeral_port(rng)
-    if second != first:
-        return second
-    port_count = EPHEMERAL_PORT_MAX - EPHEMERAL_PORT_MIN + 1
-    return EPHEMERAL_PORT_MIN + ((first - EPHEMERAL_PORT_MIN + 1) % port_count)
-
-
-def _different_ipv4(rng: random.Random, first: str) -> str:
-    second = documentation_ipv4(rng)
-    if second != first:
-        return second
-    address = ipaddress.IPv4Address(first)
-    network = next(network for network in _IPV4_DOCUMENTATION_NETWORKS if address in network)
-    offset = int(address) - int(network.network_address)
-    next_offset = 1 + (offset % (network.num_addresses - 2))
-    return str(network.network_address + next_offset)
-
-
-def _different_ipv6(rng: random.Random, first: str) -> str:
-    second = documentation_ipv6(rng)
-    if second != first:
-        return second
-    address = ipaddress.IPv6Address(first)
-    next_host = int(address) - int(_IPV6_DOCUMENTATION_NETWORK.network_address) + 1
-    if next_host >= _IPV6_DOCUMENTATION_NETWORK.num_addresses - 1:
-        next_host = 1
-    return str(ipaddress.IPv6Address(int(_IPV6_DOCUMENTATION_NETWORK.network_address) + next_host))
 
 
 if __name__ == "__main__":
